# Conflicts:
#	src/main/java/com/example/capstoneproject/controller/CvRelationController.java
#	src/main/java/com/example/capstoneproject/service/impl/CvServiceImpl.java

+++ src/main/java/com/example/capstoneproject/Dto/responses/AdminConfigurationResponse.java
@@ -12,9 +12,9 @@ import javax.persistence.*;
 @Setter
 public class AdminConfigurationResponse {
 
-    private Long vipMonthRatio;
+    private Double vipMonthRatio;
 
-    private Long vipYearRatio;
+    private Double vipYearRatio;
 
     private String apiKey;
 }
